+++ cmd/ef_tests/state/runner/levm_runner.rs
@@ -202,6 +202,7 @@ pub fn prepare_vm_for_tx<'a>(
         &tx,
         LevmCallTracer::disabled(),
         VMType::L1, // TODO: Should we run the EF tests with L2?
+        Default::default(),
     ))
 }